backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Auth
firebase-admin==6.5.0

# Database
sqlalchemy==2.0.30
aiomysql==0.2.0
cryptography==42.0.8   # needed for MariaDB SSL via aiomysql

# PDF processing
pypdf==4.2.0
pdf2docx==0.5.8
pikepdf==9.0.0
img2pdf==0.5.1
Pillow==10.3.0

# Office → PDF (calls LibreOffice CLI)
# LibreOffice must be installed on the server: apt install libreoffice

# Cloudflare R2 (S3-compatible)
boto3==1.34.112

# AI (optional — add your key to .env)
openai==1.30.1
